src/fastapi_requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
PyYAML==6.0
SQLAlchemy==1.4.46
sqlalchemy_utils==0.39.0
pycryptodome==3.16.0
requests==2.28.1
elasticsearch==8.5.3
psycopg2-binary
bs4